% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBGpoints_ly.R
\name{addBGpoints_ly}
\alias{addBGpoints_ly}
\title{addBGpoints_ly}
\usage{
addBGpoints_ly(p, data, yAxis = "y", addBG = TRUE, pointSize = 10,
  startTime = "00:00", endTime = "23:00")
}
\arguments{
\item{p}{current plot_ly plot}

\item{data}{data.frame with BG values in BG.Reading..mg.dL.}

\item{addBG}{TRUE/FALSE whether BG values should be added to current plot}

\item{pointSize}{scatter point size will be reduced by pointSize/1.5 for BG values}
}
\value{
\code{p} plot_ly interactive plot with BG values added if \code{addBG}
}
\description{
Adds BG values as scatter trace to plot_ly interactive plot \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
data<-subsetData(data,numberDays = NA,startDate = NA,endDate = NA,filterCond = "",
               startTime = "00:00", endTime = "23:00",timeStep = "hour",period = 1, 
               fromChange = TRUE,libraryPath = libraryPath)
p<-plot_ly()
addBGpoints_ly(p, data)
}
