% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBolusPoints_ly.R
\name{addBolusPoints_ly}
\alias{addBolusPoints_ly}
\title{addBolusPoints_ly}
\usage{
addBolusPoints_ly(p, data, addBolusType = "Bolus.Volume.Delivered..U.",
  pointSize = 10, startTime = "00:00", endTime = "23:00")
}
\arguments{
\item{p}{current plot_ly plot}

\item{data}{data.frame with BG values in BG.Reading..mg.dL.}

\item{addBolusType}{character string vector of Bolus columns to add as scatter points
addBolusType = c("Bolus.Volume.Delivered..U.","BWZ.Correction.Estimate..U.","BWZ.Food.Estimate..U.")}

\item{pointSize}{scatter point size}
}
\value{
\code{p} plot_ly interactive plot with BG values added if addBG
}
\description{
Adds Bolus values as scatter trace to plot_ly interactive plot \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
data<-subsetData(data,numberDays = NA,startDate = NA,endDate = NA,filterCond = "",
               startTime = "00:00", endTime = "23:00",timeStep = "hour",period = 1, 
               fromChange = TRUE,libraryPath = libraryPath)
p<-plot_ly()
addBolusPoints_ly(p,data addBolusType = c("Bolus.Volume.Delivered..U.",
                                            "BWZ.Correction.Estimate..U.","BWZ.Food.Estimate..U."))
}
