% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFasting_ly.R
\name{addFasting_ly}
\alias{addFasting_ly}
\title{addFasting_ly}
\usage{
addFasting_ly(p, data, addFasting = TRUE, addFastingAnnot = TRUE)
}
\arguments{
\item{p}{current plot_ly plot}

\item{data}{data.frame with BG values in BG.Reading..mg.dL.}

\item{addFasting}{TRUE/FALSE whether mean fasting line should be added to current plot}

\item{addFastingAnnot}{TRUE/FALSE whether mean fasting text annotation should be added to current plot}
}
\value{
\code{p} plot_ly interactive plot with mean fasting values added if \code{addFasting}
}
\description{
Adds mean fasting line that extends from earliest fasting time to latest fasting
time to plot_ly interactive plot with or without a text annotation (\code{addFastingAnnot})
printing the mean fasting value. \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
data<-subsetData(data,numberDays = NA,startDate = NA,endDate = NA,filterCond = "",
               startTime = "00:00", endTime = "23:00",timeStep = "hour",period = 1, 
               fromChange = TRUE,libraryPath = libraryPath)
p<-plot_ly()
p<-addBGpoints_ly(p, data)
addFasting_ly(p, data)
}
