% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPumpSetting_ly.R
\name{addPumpSetting_ly}
\alias{addPumpSetting_ly}
\title{addPumpSetting_ly}
\usage{
addPumpSetting_ly(p, addSetting = c("basal", "corrFactor", "carbRatio"),
  settingOverlay = FALSE, startTime = "00:00", endTime = "23:00",
  ay.list, xticks, yaxisStr, legendInset = -0.2)
}
\arguments{
\item{p}{current plot_ly plot}

\item{addSetting}{character vector of settings that should be added to current plot. Options
include 'basal', 'carbRatio', 'corrFactor', or ''}

\item{settingOverlay}{TRUE/FALSE whether settings should overlay the data or
if FALSE plot settings as subplot below data}

\item{startTime}{character string of beginning time for plot (typically startTime = "00:00)}

\item{endTime}{character string of ending time for plot (typically endTime = "23:00)}

\item{ay.list}{list of yaxis specifications for each setting,
settings are given y axis of y3, y4, y5 and ay3, ay4, ay5 respectively}

\item{xticks}{numeric vector of tickmarks for the xaxis}

\item{yaxisStr}{character string for layout of all y axes.}

\item{legendInset}{numeric value specifying how far below the plot the legend should be placed}
}
\value{
\code{p} plot_ly interactive plot with settings plotted as lines
}
\description{
Adds pump settings as line trace to plot_ly interactive plot, either overlaying or subplot \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
data<-subsetData(data,numberDays = NA,startDate = NA,endDate = NA,filterCond = "",
               startTime = "00:00", endTime = "23:00",timeStep = "hour",period = 1, 
               fromChange = TRUE,libraryPath = libraryPath)
#output as 'plot_ly'
#set parameters
addSetting<-c("basal","corrFactor","carbRatio")
settingOverlay<- FALSE
legendInset<--0.2
#format time in decimal hours
xticks.list<-xTicks(data, startTime = "00:00",endTime = "23:00",
                   timeStep = "hour",period = 1)
unPackList(lists = list(xticks.list = xticks.list),
          parentObj = list(NA)) 
#make y axis str
yaxisStr.list<-makeYaxes(addBolusType = "", addSetting,settingOverlay,
                        percentSetting = 30,addBarSub = FALSE,percentBar = NA,yTitle = "")

unPackList(lists = list(yaxisStr.list = yaxisStr.list),
          parentObj = list(NA)) 
unPackList(lists = list(ay.list = ay.list),
          parentObj = list(NA))
ay.list<-yaxisStr.list$ay.list
#get xAxis str
xaxisStr<-makeXaxis(xDomain)

#make title str
titleStr<-paste0(min(data$Date2)," -to- ",max(data$Date2))

##make layoutstr
layoutStr<-makeLayout(titleStr,xDomain,xaxisStr,yaxisStr,addGoodRange = FALSE,
                     description = "",descInset =NA)
p<-plot_ly()
#add layout
eval(parse(text = layoutStr))
p<-addBGpoints_ly(p, data)
addPumpSetting_ly(p,addSetting, settingOverlay,startTime,endTime,
                 ay.list,xticks,yaxisStr,legendInset)
}
