% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertClass.R
\name{convertClass}
\alias{convertClass}
\title{convertClass}
\usage{
convertClass(cls, t, col)
}
\arguments{
\item{cls}{character string for the desired class conversion}

\item{t}{character string for the data.frame with the column to convert}

\item{col}{charcter string for the column to convert}
}
\value{
\code{convert} vector of converted values with class equal to \code{cls}
}
\description{
Converts column classes \cr \cr
}
