% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSavedPlot.R
\name{createSavedPlot}
\alias{createSavedPlot}
\title{createSavedPlot}
\usage{
createSavedPlot(libraryPath, plotName, plotType, description, paramList,
  default = TRUE)
}
\arguments{
\item{libraryPath}{character string path to BG.library code}

\item{plotName}{character string to assign as name of saved plot}

\item{plotType}{character string indicating which plotting function to use,
options include 'plotLine_ly','summaryPlot_ly', and 'heatMap_ly'}

\item{description}{character string description of saved plot, will be added to plot as text}

\item{paramList}{list of all parameters needed to execute the plot}

\item{default}{TRUE/FALSE make this plot part of the default list of plots
allowing restoration of the original default plotList, by saving only defualt plots back
to the plotList}
}
\description{
Save a plot to BG.library/data/plotList.  These plots can be executed with
either \code{executeSavePlot()} or \code{shinyPlot()}
}
\examples{
plotName<-"meanSGheat_hist"
plotType<-"heatMap_ly"
description<-"Heat map of mean hourly SG values per day with histogram of groups."
paramList<-list(brks = c(0,50,80,150,240,300,400,500),
               brewerPallete = "RdBu",
               revPallete = TRUE,
               textCol = "black",
               tcol = "time2",
               dcol = "Date2",
               valueVar = "Sensor.Glucose..mg.dL.",
               sumFunc = "mean",
               naRemove = TRUE,
               includeTotals = TRUE,
               filterCond = "")

createSavedPlot(libraryPath, plotName,plotType,description, paramList)
}
