% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{dataImport}
\alias{dataImport}
\title{dataImport}
\usage{
dataImport(path, fileName, libraryPath)
}
\arguments{
\item{path}{character path to directory in which import file is located.}

\item{fileName}{charcter string csv file name}
}
\value{
\code{dataImport.list} list containing 2 data.frames (allData and joinData)
}
\description{
import csv file and create allData and joinData objects.
allData is pump and sensor data rbinded
joinData is a roll join of pump and sensor data on datetime.\cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
path<-"F:/BG.library_github/"
fileName<-"exampleData.csv"
#load functions
devtools::load_all(libraryPath,recompile = FALSE) 
dataImport.list<-dataImport(path,fileName,libraryPath)
}
