% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeSavedPlot.R
\name{executeSavedPlot}
\alias{executeSavedPlot}
\title{executeSavedPlot}
\usage{
executeSavedPlot(data, numberDays = NA, plotName,
  changeParam.list = NA, libraryPath)
}
\arguments{
\item{data}{data.frame to be plotted.}

\item{numberDays}{numeric value indicating number of days of data to include.  This parameter will
override \code{startDate} and \code{endDate} unless it is set to NA.  The \code{fromChange} parameter will override
all other parameters that subset the data by date.}

\item{changeParam.list}{list of parameters to change in saved plot.  Frequently changed parameters
include \code{fromChange}, \code{startDate}, and \code{endDate}}

\item{libraryPath}{character string path to BG.library code}
}
\description{
Execute a saved plot from BG.library/data/plotList \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
executeSavedPlot(data, plotName = "lineSumSens_SGper_Sett_BG", libraryPath = libraryPath)
}
