% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtables.R
\name{fixtables}
\alias{fixtables}
\title{fixtables}
\usage{
fixtables(data)
}
\arguments{
\item{data}{object from \code{readLines()} import to convert to readable data.frame.}
}
\value{
\code{data} data.frame with correct headers and dateTime class columns.
}
\description{
Reformat input tables to create readable data.frame objects. Format dateTime columns as.POSIXct.  \cr \cr
}
