% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromChangeDateRange.R
\name{fromChangeDateRange}
\alias{fromChangeDateRange}
\title{fromChangeDateRange}
\usage{
fromChangeDateRange(data, fromChange, numberDays, startDate = NA,
  endDate = NA, removeDates = NA, libraryPath)
}
\arguments{
\item{data}{data.frame to be subsetted.}

\item{fromChange}{TRUE/FALSE indicates whether data should be subset with the ealiest date
as the most recent pump settings change.  This setting overrides all other date subsetting
parameters, and must be set to \code{FALSE} to apply other parameter settings (i.e. \code{numberDays},
\code{startDate}, and \code{endDate})}

\item{numberDays}{numeric value indicating number of days of data to include.  This parameter will
override \code{startDate} and \code{endDate} unless it is set to NA.  The \code{fromChange} parameter will override
all other parameters that subset the data by date.}

\item{startDate}{Earliest date included in data.  This setting will only be applied
if \code{numberDays = NA} and \code{fromChange = FALSE}}

\item{endDate}{Latest date included in data.  This setting will only be applied
if \code{numberDays = NA} and \code{fromChange = FALSE}}

\item{removeDates}{character vector of dates in format %Y-%m-%d to remove from data}

\item{libraryPath}{character string path to BG.library code}
}
\value{
\code{data} data.frame subsetted by date
}
\description{
subset data according to date parameters. \code{fromChange = TRUE} overrides
all other date selection parameters and subsets data from the most recent pump change.
\code{!is.na(numberDays)} will overrride \code{startDate} and \code{endDate} parameters and subset data
from max(data$Date2) to the date that is \code{numberDays} away from max(data$Date2).
\code{startDate} and \code{endDate} are only applied if \code{fromChange} and \code{numberDays} are not.  Data
is subsetted with \code{min(data$Date2>=startDate)} and \code{max(data$Date2)<=endDate} \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
data<-fromChangeDateRange(data,fromChange = TRUE,numberDays = NA,
                         startDate = NA, endDate = NA, libraryPath)
range(data$Date2)
}
