% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLayout.R
\name{makeLayout}
\alias{makeLayout}
\title{makeLayout}
\usage{
makeLayout(titleStr, xDomain, xaxisStr, yaxisStr, addGoodRange,
  stackedBar = "", description, descInset)
}
\arguments{
\item{xDomain}{character string to exectute setting the domain of xaxis (format, "domain = c(0,1),")}

\item{xaxisStr}{character string to execute to set specifications for xaxis}

\item{yaxisStr}{character string to execute to set specifications for y axes}
}
\value{
\code{layoutStr} character string to execute for plot_ly layout
}
\description{
creates the character string to execute for plot_ly layout \cr \cr
}
