% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePumpSettings.R
\name{makePumpSettings}
\alias{makePumpSettings}
\title{makePumpSettings}
\usage{
makePumpSettings(libraryPath)
}
\arguments{
\item{time}{character string vector of time formatted in \code{"\%H:\%M"}}

\item{rate}{pump setting rate (i.e. basal rate, carb ratio, or correction factor)}

\item{rateName}{charcter string name of pump setting}
}
\description{
Function to create a continuous time series from pump settings vectors.  \cr \cr
}
