% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeYaxes.R
\name{makeYaxes}
\alias{makeYaxes}
\title{makeYaxes}
\usage{
makeYaxes(addBolusType, addSetting, settingOverlay, percentSetting,
  addBarSub, percentBar, yTitle)
}
\arguments{
\item{addBolusType}{character string vector of Bolus columns to add as scatter points
addBolusType = c("Bolus.Volume.Delivered..U.","BWZ.Correction.Estimate..U.","BWZ.Food.Estimate..U.")}

\item{addSetting}{character vector of settings that should be added to current plot. Options
include 'basal', 'carbRatio', 'corrFactor', or ''}

\item{settingOverlay}{TRUE/FALSE whether settings should overlay the data or
if FALSE plot settings as subplot below data}

\item{percentSetting}{numeric percentage of plotting area to dedicate to setting subplot (0-100)}

\item{addBarSub}{TRUE/FALSE indicating whether subplot of mean carb intake per hour is included}

\item{percentBar}{numeric percentage of plotting area to dedicate to carb intake bar subplot (0-100)}

\item{yTitle}{character string for left yaxis title}
}
\value{
\code{yaxisStr.list} \code{named.list(yaxisStr,ay.list, xDomain)} to execute as part of plot_ly layout
of the yaxes
}
\description{
creates lists to execute for plot_ly layout of y axes customized according to
data selection with multiple axes and scales.  \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
data<-subsetData(data,numberDays = NA,startDate = NA,endDate = NA,filterCond = "",
                startTime = "00:00", endTime = "23:00",timeStep = "hour",period = 1, 
                fromChange = TRUE,libraryPath = libraryPath)
#set paramters
addSetting<-c("basal","carbRatio")
settingOverlay<- TRUE
legendInset<--0.2

#format time in decimal hours
xticks.list<-xTicks(data, startTime = "00:00",endTime = "23:00",
                   timeStep = "hour",period = 1)
unPackList(lists = list(xticks.list = xticks.list),
          parentObj = list(NA)) 
#make y axis str
yaxisStr.list<-makeYaxes(addBolusType = "", addSetting,settingOverlay,
                        percentSetting =NA,addBarSub = FALSE,percentBar = NA,yTitle = "")

unPackList(lists = list(yaxisStr.list = yaxisStr.list),
          parentObj = list(NA)) 
unPackList(lists = list(ay.list = ay.list),
          parentObj = list(NA))
ay.list<-yaxisStr.list$ay.list
#get xAxis str
xaxisStr<-makeXaxis(xDomain)

#make title str
titleStr<-paste0(min(data$Date2)," -to- ",max(data$Date2))

##make layoutstr
layoutStr<-makeLayout(titleStr,xDomain,xaxisStr,yaxisStr,addGoodRange = FALSE,
                     description = "",descInset =NA)
p<-plot_ly()
#add layout
eval(parse(text = layoutStr))
p<-addBGpoints_ly(p, data)
addPumpSetting_ly(p,addSetting, settingOverlay,startTime,endTime,
                 ay.list,xticks,yaxisStr,legendInset)
}
