% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeYaxesSetting.R
\name{makeYaxesSetting}
\alias{makeYaxesSetting}
\title{makeYaxesSetting}
\usage{
makeYaxesSetting(addSetting, settingOverlay, addBolusType, barSubPlot,
  allPosition, position, numberAxes, yDomain2, yDomain3, yaxisStr)
}
\arguments{
\item{addSetting}{character vector of settings to plot c("basal,"corrFactor","carbRatio")}

\item{settingOverlay}{TRUE/FALSE whether or not settings overlay main plot}

\item{addBolusType}{character vector of types of bolus insulin delivered (addBolusType = "" for none)}

\item{barSubPlot}{TRUE/FALSE whether barsubplot is to be included}

\item{allPosition}{numeric vector of positions of all y axes}

\item{position}{numeric value for y axes position (0-1 along x axes)}

\item{numberAxes}{numeric value for current number of y axes}

\item{yDomain2}{numeric vector for y domain of second subplot}

\item{yDomain3}{numeric vector for y domain of third subplot}

\item{yaxisStr}{current character string for layout of all y axes}
}
\value{
\code{settingAxis.list} named.list(yaxisStr, position, numberAxes)
}
\description{
Create layout lists for yaxes associated with pump settings \cr \cr
}
