% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeYaxesSummary.R
\name{makeYaxesSummary}
\alias{makeYaxesSummary}
\title{makeYaxesSummary}
\usage{
makeYaxesSummary(addSetting, settingOverlay, percentSetting, barSubPlot,
  addBG, initYrange, yTitle)
}
\arguments{
\item{addSetting}{character vector of settings to plot c("basal,"corrFactor","carbRatio")}

\item{settingOverlay}{TRUE/FALSE whether or not settings overlay main plot}

\item{percentSetting}{numeric value (0-100) for percentage of total plot to dedicate to setting subplot}

\item{barSubPlot}{TRUE/FALSE whether barsubplot is to be included}

\item{initYrange}{numeric vector for range of initial y axis on the left}
}
\value{
\code{yaxisStr.list} named.list(yaxisStr,ay.list, xDomain)
}
\description{
Set up y axes for barplot as main plot \cr \cr
}
