% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeYdomain.R
\name{makeYdomain}
\alias{makeYdomain}
\title{makeYdomain}
\usage{
makeYdomain(percentSetting, percentBar, addSetting, settingOverlay,
  addBarSub)
}
\arguments{
\item{percentSetting}{numeric value (0-100) for percentage of total plot to dedicate to setting subplot}

\item{percentBar}{numeric value (0-100) for percentage of total plot to dedicate to bar subplot}

\item{addSetting}{character vector of settings to plot c("basal,"corrFactor","carbRatio")}

\item{settingOverlay}{TRUE/FALSE whether or not settings overlay main plot}

\item{addBarSub}{TRUE/FALSE whether barsubplot is to be included}
}
\value{
\code{domain.list} named.list(percentSetting, percentBar, yDomain1, yDomain2, yDomain3)
}
\description{
Set up y domains 1,2, and 3 for subplots \cr \cr
}
\examples{
#set yDomain
domain.list<-makeYdomain(percentSetting = 30,percentBar = NA,addSetting = c("basal","carbRatio"),
                        settingOverlay = TRUE,addBarSub = FALSE)
}
