% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.list.R
\name{named.list}
\alias{named.list}
\title{named.list}
\usage{
named.list(...)
}
\value{
named list object
}
\description{
Creates a list object by specifying unquoted variable names in
the argument list. \cr \cr
}
\examples{
x<-list(x = c(1,2,3), x2 = c(2,3,4))
y<-c("a","b","c")
named.list(x,y)
}
