% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLine_ly.R
\name{plotLine_ly}
\alias{plotLine_ly}
\title{plotLine_ly}
\usage{
plotLine_ly(data, scatterOnly = FALSE, addBG = TRUE,
  addPercentBG = c("low", "good", "high", "very high"),
  addPercentType = "BG.Reading..mg.dL.",
  addBolusType = "Bolus.Volume.Delivered..U.",
  plotSummary = "Sensor.Glucose..mg.dL.", addSetting = "",
  settingOverlay = FALSE, percentSetting = 30, addBarSub = TRUE,
  percentBar = 30, addGoodRange = TRUE, addFasting = TRUE,
  addFastingAnnot = TRUE, fromChange = TRUE, numberDays = NA,
  startDate = NA, endDate = NA, removeDates = NA,
  startTime = "00:00", endTime = "23:00", timeStep = "hour",
  period = 1, filterCond = "", colorPalleteDaily = "rainbow",
  pointSize = 10, legendInset = -0.2, description = "",
  descInset = -0.15, libraryPath)
}
\arguments{
\item{data}{data.frame with BG values in BG.Reading..mg.dL. and SG values in Sensor.Glucose..mg.dL.}

\item{scatterOnly}{TRUE/FALSE indicating whether only points with no lines are to be plotted}

\item{addBG}{TRUE/FALSE whether BG values should be added to current plot}

\item{addPercentBG}{character vector of groups to include (c("low","good","high","very high"))}

\item{addPercentType}{character string of column name with values to group (i.e. "BG.Reading..mg.dL.")}

\item{addBolusType}{character string vector of Bolus columns to add as scatter points
addBolusType = c("Bolus.Volume.Delivered..U.","BWZ.Correction.Estimate..U.","BWZ.Food.Estimate..U.")}

\item{plotSummary}{glucose values to summarise outputing min, mean, and max lines to plot
common options are 'BG.Reading..mg.dL.' or 'Sensor.Glucose..mg.dL.'}

\item{addSetting}{character vector of settings that should be added to current plot. Options
include 'basal', 'carbRatio', 'corrFactor', or ''}

\item{settingOverlay}{TRUE/FALSE whether settings should overlay the data or
if FALSE plot settings as subplot below data}

\item{percentSetting}{numeric percentage of plotting area to dedicate to setting subplot (0-100)}

\item{addBarSub}{TRUE/FALSE indicating whether subplot of mean carb intake per hour is included}

\item{percentBar}{numeric percentage of plotting area to dedicate to carb intake bar subplot (0-100)}

\item{addGoodRange}{TRUE/FALSE indicating whether shaded polygon for good BG range is plotted}

\item{addFasting}{TRUE/FALSE whether mean fasting line should be added to current plot}

\item{addFastingAnnot}{TRUE/FALSE whether mean fasting text annotation should be added to current plot}

\item{fromChange}{TRUE/FALSE indicates whether data should be subset with the ealiest date
as the most recent pump settings change.  This setting overrides all other date subsetting
parameters, and must be set to \code{FALSE} to apply other parameter settings (i.e. \code{numberDays},
\code{startDate}, and \code{endDate})}

\item{numberDays}{numeric value indicating number of days of data to include.  This parameter will
override \code{startDate} and \code{endDate} unless it is set to NA.  The \code{fromChange} parameter will override
all other parameters that subset the data by date.}

\item{startDate}{Earliest date included in data.  This setting will only be applied
if \code{numberDays = NA} and \code{fromChange = FALSE}}

\item{endDate}{Latest date included in data.  This setting will only be applied
if \code{numberDays = NA} and \code{fromChange = FALSE}}

\item{startTime}{character string of beginning time for plot (typically startTime = "00:00)}

\item{endTime}{character string of ending time for plot (typically endTime = "23:00)}

\item{timeStep}{character string indicating the time step to aggregate data, possible values
include c("hour","day")}

\item{period}{numeric value indicating number of \code{timeSteps} to aggregate into single step
for example : \code{timeStep = 'hour'}  and \code{period = 3} outputs plots with tick marks every 3 hours.}

\item{filterCond}{character string of R syntax to be applied to filter the data,
example \code{data[data$BG.Reading..mg.dL.>150 & !is.na(data$BG.Reading..mg.dL.),]}}

\item{colorPalleteDaily}{character string color pallete for daily sensor lines}

\item{pointSize}{scatter point size will be reduced by pointSize/1.5 for BG values}

\item{legendInset}{numeric value specifying how far below the plot the legend should be placed}

\item{description}{character string plot description to be output as part of plot}

\item{descInset}{numeric value to place description below plot}

\item{libraryPath}{character string path to BG.library code}
}
\description{
Creates interactive plotly scatter and line plots \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
#line plot
plotLine_ly(data,addBolusType = "Bolus.Volume.Delivered..U.",
           numberDays = 5,libraryPath = libraryPath)
}
