% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSensor.R
\name{plotSensor}
\alias{plotSensor}
\title{plotSensor}
\usage{
plotSensor(data, numberDays, colorPallete = "rainbow", addBG = TRUE,
  addBolus = TRUE, addBolusType = "Bolus.Volume.Delivered..U.",
  sensorSummary = FALSE, addBasal, addCorrfactor, addCarbRatio)
}
\arguments{
\item{data}{the data.frame containing both pump and sensor data to be used for plotting}

\item{numberDays}{number of days to plot}

\item{colorPallete}{charcter string for color pallete used to plot daily sensor data lines}

\item{addBG}{TRUE/FALSE indicating whether BG values should be plotted}

\item{addBolus}{TRUE/FALSE indicating whether insulin bolus data should be plotted}

\item{addBolusType}{vector of character strings for column names of insulin bolus columns to be plotted}

\item{sensorSummary}{TRUE/FALSE indicating whether daily sensor values should be plotted as min, mean, max}

\item{addBasal}{TRUE/FALSE indicating whether basal settings should be plotted}

\item{addCorrfactor}{TRUE/FALSE indicating whether correction factor settings should be plotted}

\item{addCarbRatio}{TRUE/FALSE indicating whether carb ratio settings should be plotted}
}
\description{
Function to plot sensor and pump data, with options to summarize sensor data, and add pump settings. \cr \cr
}
