% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDupBGs.R
\name{removeDupBGs}
\alias{removeDupBGs}
\title{removeDupBGs}
\usage{
removeDupBGs(data)
}
\arguments{
\item{data}{data.frame with BG values in BG.Reading..mg.dL.}
}
\value{
\code{data} data.frame with duplicates removed
}
\description{
remove duplicate BG values within 10 minute intervals
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
#readBG
data<-read.csv(filePath, header = FALSE)

#get pump table
pumpStartRow<-which(data[,1]=="Index")[1]
pumpEndRow<-which(data[,1]=="-------" & as.numeric(rownames(data))>pumpStartRow)-1
pumpData<-data[pumpStartRow:pumpEndRow,]

#fix tables
pumpData<-fixtables(pumpData,libraryPath)
nrow(pumpData)

#remove duplicates
pumpData<-removeDupBGs(pumpData)
nrow(pumpData)
}
