% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceTitle.R
\name{replaceTitle}
\alias{replaceTitle}
\title{replaceTitle}
\usage{
replaceTitle(libraryPath, reportTitle, reportName)
}
\arguments{
\item{libraryPath}{character string path to BG.library code}

\item{reportTitle}{character string title to use for RMD document}

\item{reportName}{character string indicating the name of the report to
edit (i.e. "BG_report","historySeqOut")}
}
\description{
Replace title of RMD reports with custom title  \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
reportTitle<-"This is my new BG_report"
replaceTitle(libraryPath, reportTitle,reportName = "BG_report")
}
