% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTimeStep.R
\name{setTimeStep}
\alias{setTimeStep}
\title{setTimeStep}
\usage{
setTimeStep(data, startTime = "00:00", endTime = "23:00", timeStep,
  period)
}
\arguments{
\item{data}{data.frame with dateTime, time2, Date2 columns}

\item{startTime}{character string of beginning time for plot (typically startTime = "00:00)}

\item{endTime}{character string of ending time for plot (typically endTime = "23:00)}

\item{timeStep}{character string indicating the time step to aggregate data, possible values
include c("hour","day")}

\item{period}{numeric value indicating number of \code{timeSteps} to aggregate into single step
for example : \code{timeStep = 'hour'}  and \code{period = 3} outputs plots with tick marks every 3 hours.}
}
\value{
\code{data} data.frame with new aggregated timeSteps
}
\description{
Reformat dateTime columns creating new columns with custom \code{timeStep}
and \code{period}  \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
unique(data$hour)
data<-setTimeStep(data,timeStep = "hour", period = 3)
unique(data$hour)
}
