% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPlot.R
\name{shinyPlot}
\alias{shinyPlot}
\title{shinyPlot}
\usage{
shinyPlot(libraryPath, filePath)
}
\arguments{
\item{libraryPath}{character string path to BG.library code}

\item{filePath}{character path to csv import file}
}
\description{
Shiny application for generating interactive plotly plots, BG_report, and
historical sequence plots with user-friendly shiny user inputs
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
shinyPlot(libraryPath, filePath)
}
