% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetData.R
\name{subsetData}
\alias{subsetData}
\title{subsetData}
\usage{
subsetData(data, fromChange = TRUE, numberDays = NA, startDate = NA,
  endDate = NA, removeDates = NA, startTime = "00:00",
  endTime = "23:00", timeStep = "hour", period = 1,
  filterCond = "", libraryPath)
}
\arguments{
\item{data}{data.frame to be subsetted.}

\item{fromChange}{TRUE/FALSE indicates whether data should be subset with the ealiest date
as the most recent pump settings change.  This setting overrides all other date subsetting
parameters, and must be set to \code{FALSE} to apply other parameter settings (i.e. \code{numberDays},
\code{startDate}, and \code{endDate})}

\item{numberDays}{numeric value indicating number of days of data to include.  This parameter will
override \code{startDate} and \code{endDate} unless it is set to NA.  The \code{fromChange} parameter will override
all other parameters that subset the data by date.}

\item{startDate}{Earliest date included in data.  This setting will only be applied
if \code{numberDays = NA} and \code{fromChange = FALSE}}

\item{endDate}{Latest date included in data.  This setting will only be applied
if \code{numberDays = NA} and \code{fromChange = FALSE}}

\item{removeDates}{character vector of dates in format %Y-%m-%d to remove from data}

\item{startTime}{character string of beginning time for plot (typically startTime = "00:00)}

\item{endTime}{character string of ending time for plot (typically endTime = "23:00)}

\item{timeStep}{character string indicating the time step to aggregate data, possible values
include c("hour","day")}

\item{period}{numeric value indicating number of \code{timeSteps} to aggregate into single step
for example : \code{timeStep = 'hour'}  and \code{period = 3} outputs plots with tick marks every 3 hours.}

\item{filterCond}{character string of R syntax to be applied to filter the data,
example \code{data[data$BG.Reading..mg.dL.>150 & !is.na(data$BG.Reading..mg.dL.),]}}

\item{libraryPath}{character string path to BG.library code}
}
\value{
\code{data} data.frame subsetted by date
}
\description{
subset data according to date and time parameters taking into account time step.
\code{fromChange = TRUE} overrides all other date selection parameters and subsets data from the
most recent pump change.  \code{!is.na(numberDays)} will overrride \code{startDate} and \code{endDate} parameters
and subset data from max(data$Date2) to the date that is \code{numberDays} away from max(data$Date2).
\code{startDate} and \code{endDate} are only applied if \code{fromChange} and \code{numberDays} are not.  Data
is subsetted with \code{min(data$Date2>=startDate)} and \code{max(data$Date2)<=endDate} \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
range(data$Date2)
data<-subsetData(data = data,libraryPath = libraryPath)
range(data$Date2)
}
