% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetSetting.R
\name{subsetSetting}
\alias{subsetSetting}
\title{subsetData}
\usage{
subsetSetting(data, libraryPath)
}
\arguments{
\item{data}{data.frame to be subsetted.}

\item{libraryPath}{character string path to BG.library code}
}
\value{
\code{pumpSettings.list} named list of all pump settings
}
\description{
subset settings showing only settings <= max(data$Date2) \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
pumpSettings.list<-makePumpSettings(libraryPath)
print(names(pumpSettings.list$basal))
pumpSettings.list<-subsetSetting(data = data,libraryPath = libraryPath)
print(names(pumpSettings.list$basal))
}
