% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeData.R
\name{summarizeData}
\alias{summarizeData}
\title{summarizeData}
\usage{
summarizeData(data, colName, sumFuncs = "min, mean, max, sd",
  fromChange = TRUE, numberDays = NA, startDate = NA, endDate = NA,
  removeDates = NA, startTime = "00:00", endTime = "23:00",
  timeStep = "hour", period = 1, filterCond = "", libraryPath)
}
\arguments{
\item{data}{data.frame to be summarized}

\item{colName}{character string name of column to summarize}

\item{sumFuncs}{character vector of aggregate functions to apply to \code{colName}}

\item{fromChange}{TRUE/FALSE indicates whether data should be subset with the ealiest date
as the most recent pump settings change.  This setting overrides all other date subsetting
parameters, and must be set to \code{FALSE} to apply other parameter settings (i.e. \code{numberDays},
\code{startDate}, and \code{endDate})}

\item{numberDays}{numeric value indicating number of days of data to include.  This parameter will
override \code{startDate} and \code{endDate} unless it is set to NA.  The \code{fromChange} parameter will override
all other parameters that subset the data by date.}

\item{startDate}{Earliest date included in data.  This setting will only be applied
if \code{numberDays = NA} and \code{fromChange = FALSE}}

\item{endDate}{Latest date included in data.  This setting will only be applied
if \code{numberDays = NA} and \code{fromChange = FALSE}}

\item{removeDates}{character vector of dates in format %Y-%m-%d to remove from data}

\item{startTime}{character string of beginning time for plot (typically startTime = "00:00)}

\item{endTime}{character string of ending time for plot (typically endTime = "23:00)}

\item{timeStep}{character string indicating the time step to aggregate data, possible values
include c("hour","day")}

\item{period}{numeric value indicating number of \code{timeSteps} to aggregate into single step
for example : \code{timeStep = 'hour'}  and \code{period = 3} outputs plots with tick marks every 3 hours.}

\item{filterCond}{character string of R syntax to be applied to filter the data,
example \code{data[data$BG.Reading..mg.dL.>150 & !is.na(data$BG.Reading..mg.dL.),]}}

\item{libraryPath}{character string path to BG.library code}
}
\value{
\code{data} data.frame of \code{colName} summarized by time and date
}
\description{
summarize data column (\code{colName}) by \code{sumFunc} and output data.frame \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
BGvalue_Summary<-summarizeData(data, colName = "BG.Reading..mg.dL.", 
                              libraryPath = libraryPath)
}
