% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryLinePlot_ly.R
\name{summaryLinePlot_ly}
\alias{summaryLinePlot_ly}
\title{summaryLinePlot_ly}
\usage{
summaryLinePlot_ly(p, data, plotSummary)
}
\arguments{
\item{p}{current plot_ly plot}

\item{data}{data.frame with data to use for summary plot lines}

\item{plotSummary}{glucose values to summarise outputing min, mean, and max lines to plot
common options are 'BG.Reading..mg.dL.' or 'Sensor.Glucose..mg.dL.'}
}
\value{
\code{p} plotly plot object
}
\description{
Adds summary glucose value lines (min, mean, and max) to plotly plot for column
defined in \code{plotSummary} \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
data<-subsetData(data,numberDays = NA,startDate = NA,endDate = NA,filterCond = "",
                startTime = "00:00", endTime = "23:00",timeStep = "hour",period = 1, 
                fromChange = TRUE,libraryPath = libraryPath)
p<-plot_ly()
p<-addBGpoints_ly(p, data)
summaryLinePlot_ly(p, data, plotSummary = "Sensor.Glucose..mg.dL.")
}
