% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPlot_ly.R
\name{summaryPlot_ly}
\alias{summaryPlot_ly}
\title{summaryPlot_ly}
\usage{
summaryPlot_ly(p, data, ayCarb, addBarSub = FALSE, boxBar = "bar",
  numberDays, filterCond = "", startDate = NA, endDate = NA,
  removeDates = NA, startTime = "00:00", endTime = "23:00",
  timeStep = "hour", period = 1, fromChange = TRUE, libraryPath,
  plotSummary, sumFunc = "length", stackedBar = "", uniqueDT = TRUE,
  replaceNAs = TRUE, ignoreNAs = FALSE, addBG = TRUE,
  pointSize = 10, addSetting = "", settingOverlay = FALSE,
  percentSetting = 30, legendInset = -0.2, description = "",
  descInset = -0.15)
}
\arguments{
\item{p}{current plot_ly plot}

\item{data}{data.frame with BG values in BG.Reading..mg.dL.}

\item{ayCarb}{list of y axis specifications for carb barplot}

\item{barSubPlot}{TRUE/FALSE whether subplot should be added to main plot}
}
\value{
\code{p} plot_ly interactive plot
}
\description{
generate bar and box summary plots and carb subbarplot  \cr \cr
}
