% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{dataImport}
\alias{dataImport}
\title{dataImport}
\usage{
dataImport(filePath, libraryPath)
}
\arguments{
\item{filePath}{character path to csv import file}

\item{libraryPath}{character string path to BG.library code}
}
\value{
\code{dataImport.list} list containing 2 data.frames (allData and joinData)
}
\description{
import csv file and create allData and joinData objects.
allData is pump and sensor data rbinded
joinData is a roll join of pump and sensor data on datetime.\cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
}
