% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixtables.R
\name{fixtables}
\alias{fixtables}
\title{fixtables}
\usage{
fixtables(data, libraryPath)
}
\arguments{
\item{data}{object from \code{readLines()} import to convert to readable data.frame.}
}
\value{
\code{data} data.frame with correct headers and dateTime class columns.
}
\description{
Reformat input tables to create readable data.frame objects. Format dateTime columns as.POSIXct.  \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
#readBG
data<-read.csv(filePath, header = FALSE)

#get pump table
pumpStartRow<-which(data[,1]=="Index")[1]
pumpEndRow<-which(data[,1]=="-------" & as.numeric(rownames(data))>pumpStartRow)-1
pumpData<-data[pumpStartRow:pumpEndRow,]

#fix tables
pumpData<-fixtables(pumpData,libraryPath)
}
