% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePumpSettings.R
\name{makePumpSettings}
\alias{makePumpSettings}
\title{makePumpSettings}
\usage{
makePumpSettings(libraryPath)
}
\arguments{
\item{libraryPath}{character string path to BG.library code}
}
\description{
Function to create a continuous time series from pump settings csvs  \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
pumpSettings.list<-makePumpSettings(libraryPath)
}
