% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeYaxesSetting.R
\name{makeYaxesSetting}
\alias{makeYaxesSetting}
\title{makeYaxesSetting}
\usage{
makeYaxesSetting(addSetting, settingOverlay, addBolusType, addBarSub,
  allPosition, position, numberAxes, yDomain2, yDomain3, yaxisStr)
}
\arguments{
\item{addSetting}{character vector of settings to plot c("basal,"corrFactor","carbRatio")}

\item{settingOverlay}{TRUE/FALSE whether or not settings overlay main plot}

\item{addBolusType}{character vector of types of bolus insulin delivered (addBolusType = "" for none)}

\item{addBarSub}{TRUE/FALSE whether barsubplot is to be included}

\item{allPosition}{numeric vector of positions of all y axes}

\item{position}{numeric value for y axes position (0-1 along x axes)}

\item{numberAxes}{numeric value for current number of y axes}

\item{yDomain2}{numeric vector for y domain of second subplot}

\item{yDomain3}{numeric vector for y domain of third subplot}

\item{yaxisStr}{current character string for layout of all y axes}
}
\value{
\code{settingAxis.list} named.list(yaxisStr, position, numberAxes)
}
\description{
Create layout lists for yaxes associated with pump settings \cr \cr
}
\examples{
#set paramters
addSetting<-c("basal","carbRatio")
settingOverlay<- TRUE
legendInset<--0.2
percentSetting<-30
percentBar<-NA
addBarSub<-FALSE
addBolusType<-""
yTitle<-""

#set yDomain
domain.list<-makeYdomain(percentSetting,percentBar,addSetting,settingOverlay,addBarSub)
unPackList(lists = list(domain.list = domain.list),
          parentObj = list(NA)) 

#get number axes
if (addSetting[1]!=""){
 numberAxes<-length(addSetting[addSetting!=""])
}else{
 numberAxes<-1
}

#set position
position<-1
allPosition<-position

#itialize plotly plot y axis
yaxisStr<-paste0("yaxis = list(range = c(0,450), ticks = 'outside', zeroline = FALSE,showline = TRUE,
                              title = '",yTitle, "',
                              domain = c(",paste(yDomain1,collapse = ","),"))")
#create y axes for pump settings
settingAxis.list<-makeYaxesSetting(addSetting, settingOverlay, addBolusType,addBarSub,
                                  allPosition, position,numberAxes,yDomain2,yDomain3,yaxisStr)
}
