% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeYaxesSummary.R
\name{makeYaxesSummary}
\alias{makeYaxesSummary}
\title{makeYaxesSummary}
\usage{
makeYaxesSummary(addSetting, settingOverlay, percentSetting, addBarSub,
  addBG, initYrange, yTitle)
}
\arguments{
\item{addSetting}{character vector of settings to plot c("basal,"corrFactor","carbRatio")}

\item{settingOverlay}{TRUE/FALSE whether or not settings overlay main plot}

\item{percentSetting}{numeric value (0-100) for percentage of total plot to dedicate to setting subplot}

\item{addBarSub}{TRUE/FALSE whether barsubplot is to be included}

\item{addBG}{TRUE/FALSE whether BG values should be added to current plot}

\item{initYrange}{numeric vector for range of initial y axis on the left}

\item{yTitle}{character string for left yaxis title}
}
\value{
\code{yaxisStr.list} named.list(yaxisStr,ay.list, xDomain)
}
\description{
Set up y axes for summaryPlot_ly as main plot \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
data<-subsetData(data,numberDays = NA,startDate = NA,endDate = NA,filterCond = "",
                startTime = "00:00", endTime = "23:00",timeStep = "hour",period = 1, 
                fromChange = TRUE,libraryPath = libraryPath)

#set parameters
boxBar<-"bar"
sumFunc<-"mean"
plotSummary<-"BG.Reading..mg.dL."
addSetting<-c("basal","carbRatio")
settingOverlay<-FALSE
percentSetting<-30
addBarSub<-FALSE
addBG<-FALSE

#subset settings
pumpSettings.list<-subsetSetting(data,libraryPath)
unPackList(lists = list(pumpSettings.list = pumpSettings.list),
          parentObj = list(NA)) 

#set up data for summaryPlot_ly
dataOrig<-data
data$temp<-eval(parse(text = paste0("data$",plotSummary)))
data<-data[c("hour","temp")]
dataFormat<-data

#set y axis parameters
initYrange<-c(0,450)
yTitle<-ifelse(sumFunc!="length",paste0(sumFunc,"_",plotSummary),paste0("number_",plotSummary))

#format time in decimal hours for xaxis tick marks
xticks.list<-xTicks(data = dataOrig, startTime = "00:00", endTime = "23:00",timeStep = "hour",period = 1)
unPackList(lists = list(xticks.list = xticks.list),
          parentObj = list(NA)) 

#get yaxis code string
yaxisStr.list<-makeYaxesSummary(addSetting, settingOverlay, percentSetting,addBarSub,addBG,
                               initYrange,yTitle)

unPackList(lists = list(yaxisStr.list = yaxisStr.list),
          parentObj = list(NA)) 
unPackList(lists = list(ay.list = ay.list),
          parentObj = list(NA))
ay.list<-yaxisStr.list$ay.list

#get xAxis str
xaxisStr<-makeXaxis(xDomain)

#make title str
titleStr<-paste0(min(data$Date2)," -to- ",max(data$Date2))

##make layoutstr
layoutStr<-makeLayout(titleStr,xDomain,xaxisStr,yaxisStr,addGoodRange = FALSE,stackedBar = "",
                     description = "",descInset = NA)
#initialize plot
p<-plot_ly()
#add layout
eval(parse(text = layoutStr))
#get formatted data
data<-dataFormat

#summarize data
sumString<-paste0("as.data.frame(data \%>\% group_by(hour) \%>\% summarise_all(funs(",
                 sumFunc,"),na.rm = TRUE))")
data<-eval(parse(text = sumString))

#add bars to plot
p \%>\% add_trace(data = data, x = ~hour, y = ~temp,type = 'bar', 
               name = paste0(sumFunc,"_",plotSummary)) 
}
