% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyTime.R
\name{prettyTime}
\alias{prettyTime}
\title{prettyTime}
\usage{
prettyTime(data, timeCol)
}
\value{
data.frame with new formatted date and time columns
}
\description{
Formats date and time columns for processing from a "POSIXct" "POSIXt"
class dateTime column \cr \cr
}
\examples{
libraryPath<-"F:/BG.library_github/BG.library/"
filePath<-"F:/BG.library_github/exampleData.csv"
dataImport.list<-dataImport(filePath,libraryPath)
data<-dataImport.list$allData
data<-prettyTime(data,"dateTime")
}
